Change Database to allow multiple coppies of a book, restructing middlewares and bug fixes

+++ src/controllers/adminControllers.js
@@ -70,14 +70,24 @@ async function saveBookEditChanges(req,res,buid){
     
     const inp_name=req.body.name;
     const inp_description=req.body.description;
+    const inp_quantity=req.body.qty;
     if (inp_description.length>2000){
         return res.send("Can't update details: Description too long!")
     }
     if (inp_name.length>50){
         return res.send("Can't update details: Name too long!")
     }
-    let result= await dbQuery(`UPDATE BOOKS SET NAME=? WHERE BUID=?`,[inp_name,buid]);
+    let result= await dbQuery("SELECT * FROM BOOKS WHERE BUID=?;",[buid])
+    const checked_out=result[0]["TOTAL"]-result[0]["CHECKIN"]
+    if(inp_quantity<checked_out){
+        return res.status(403).send("Can't decrease quantity under a certain value")
+    }
+    let decrease=result[0]["TOTAL"]-inp_quantity
+    let decrease_checkin= result[0]["CHECKIN"]-decrease
+    result= await dbQuery(`UPDATE BOOKS SET NAME=? WHERE BUID=?`,[inp_name,buid]);
     result=await dbQuery(`UPDATE BOOKS SET DESCRIPTION=? WHERE BUID=?`,[inp_description,buid]);
+    result=await dbQuery(`UPDATE BOOKS SET TOTAL=? WHERE BUID=?`,[inp_quantity,buid]);
+    result=await dbQuery(`UPDATE BOOKS SET CHECKIN=? WHERE BUID=?`,[decrease_checkin,buid]);
     res.status(200).redirect(`/books/${buid}`);
 }
 
@@ -85,20 +95,29 @@ async function saveBookEditChanges(req,res,buid){
 async function newBook(req,res){
     const inp_name=req.body.name;
     const inp_description=req.body.description;
+    const inp_quantity=req.body.qty;
     if (inp_description.length>2000){
         return res.send("Can't update details: Description too long!")
-    }
+    } else if (inp_description.trim().length==0){
+        return res.send("Can't update details: Description too short!")}
     if (inp_name.length>50){
         return res.send("Can't update details: Name too long!")
-    }
-    let result= await dbQuery(`INSERT INTO BOOKS VALUES(NULL,?,?,1)`,[inp_name,inp_description]);
+    } else if (inp_name.trim().length==0){
+        return res.send("Can't update details: Name too short!")}
+    
+    let result= await dbQuery(`INSERT INTO BOOKS VALUES(NULL,?,?,?,?)`,[inp_name.trim(),inp_description.trim(),inp_quantity,inp_quantity]);
+    
     res.status(200).redirect(`/books`);
 }
 
 //admin delete book
 async function deleteBook(req,res,buid){
-    
-    let result= await dbQuery(`DELETE FROM BOOKS WHERE BUID=?`,[buid]);
+    let result= await dbQuery(`Select * FROM BOOKS WHERE BUID=?`,[buid]);
+    if (result[0]["CHECKIN"]!=result[0]["TOTAL"]){   // cant delete book if all books are not checked in
+        return res.redirect(`/books/${buid}`)
+    }
+    req.cantdelete=false
+    let result2= await dbQuery(`DELETE FROM BOOKS WHERE BUID=?`,[buid]);
     res.status(200).redirect(`/books`);
 }
 
@@ -107,41 +126,49 @@ async function approve(req,res,uuid,buid){
     // checkin approval reqires unavailablity of book and consecutive checkin entry in pending_requests     
     
     const result=await dbQuery("SELECT * FROM BOOKS WHERE BUID=?",[buid])
+    let result2=await dbQuery("SELECT * FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?",[uuid,buid])
     
     if (result.length!=1){
-        return res.send("Incorrect BUID parameter!");
-    }
-    if (result[0]["CHECKIN"]==0){         // checkin approval
-        let result2 = await dbQuery("SELECT * FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid]);
-        if (result2.length==1 && result2[0]["TYPE"]==1){
-            result2 = await dbQuery("DELETE FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid])
-            result2 = await dbQuery("SELECT * FROM BORROWING_HISTORY WHERE UUID=? AND BUID=? AND CHECKIN_DATE IS NULL ",[uuid,buid])
-            const d=new Date();
-            let date=d.getFullYear()+'-'+(d.getMonth()+1)+'-'+(d.getDate());
-            result2 = await dbQuery("UPDATE BORROWING_HISTORY SET CHECKIN_DATE=? WHERE UUID=? AND BUID=? AND CHECKOUT_DATE=? ",[date,uuid,buid,result2[0]["CHECKOUT_DATE"]])
-            result2= await dbQuery("UPDATE BOOKS SET CHECKIN=1 WHERE BUID=?;",[buid])
-        }
-    }
-    if (result[0]["CHECKIN"]==1){       // checkout approval
-        let result2 = await dbQuery("SELECT * FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid]);
-        if (result2.length==1 && result2[0]["TYPE"]==0){
-            result2 = await dbQuery("DELETE FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid])
-            const d=new Date();
-            let date=d.getFullYear()+'-'+(d.getMonth()+1)+'-'+(d.getDate());
-            result2 = await dbQuery("INSERT INTO BORROWING_HISTORY VALUES(?,?,?,NULL)",[uuid,buid,date])
-            result2= await dbQuery("UPDATE BOOKS SET CHECKIN=0 WHERE BUID=?;",[buid])
-        }
+        return res.status(400).send("Incorrect BUID parameter!");
     }
 
-    const role=req.user.role;
+    if(result2.length!=1){
+        return res.status(400).send("Pending Request not found!");
 
-    if (role==="admin"){
-        return res.redirect("/")
     }
+    
+    
+    if (result2[0]["TYPE"]==1){     //checkin approval
+
+        let result3 = await dbQuery("DELETE FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid])
+        result3 = await dbQuery("SELECT * FROM BORROWING_HISTORY WHERE UUID=? AND BUID=? AND CHECKIN_DATE IS NULL ",[uuid,buid])
+        const d=new Date();
+        let date=d.getFullYear()+'-'+(d.getMonth()+1)+'-'+(d.getDate());
+        result3 = await dbQuery("UPDATE BORROWING_HISTORY SET CHECKIN_DATE=? WHERE UUID=? AND BUID=? AND CHECKOUT_DATE=? ",[date,uuid,buid,result3[0]["CHECKOUT_DATE"]]);
+        result3= await dbQuery("UPDATE BOOKS SET CHECKIN=? WHERE BUID=?;",[(result[0]["CHECKIN"]+1),buid])
+        
+    }
+    
+    if (result2[0]["TYPE"]==0 && result[0]["CHECKIN"]>=1){   //checkout approval 
+        let  result3= await dbQuery("DELETE FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid])
+        const d=new Date();
+        let date=d.getFullYear()+'-'+(d.getMonth()+1)+'-'+(d.getDate());
+        result3 = await dbQuery("INSERT INTO BORROWING_HISTORY VALUES(?,?,?,NULL)",[uuid,buid,date])
+        result3= await dbQuery("UPDATE BOOKS SET CHECKIN=? WHERE BUID=?;",[(result[0]["CHECKIN"]-1),buid])
+        
+    }
+
+    result2=await dbQuery("SELECT * FROM USER WHERE UUID=?;",[uuid]);
+
+    if (result2[0]["ROLE"]==="admin"){
+        return res.redirect("/");
+    }
+
     res.redirect("/pending")
+    
 }
 
-async function deny(req,res,uuid,buid){////////////////////////////////////////////////////////////////////////////////
+async function deny(req,res,uuid,buid){
     let result = await dbQuery("DELETE FROM PENDING_REQUESTS WHERE UUID=? AND BUID=?;",[uuid,buid])
     res.redirect("/pending")
 }
